refactor(cli): migrate logs commands from legacy api client to domain module
Replace the last 2 consumers of the legacy ApiClient class with the
domain modules pattern, then delete api-client.ts (772 lines).

- Create lib/api/domains/logs.ts with 5 functions: getSystemLog,
  getMetrics, getAgentEvents, getNetworkLogs, searchLogs
- Update commands/logs/index.ts and search.ts to use domain module
- Normalize error handling from throw Error to handleError()
- Delete lib/api/api-client.ts entirely
- Update knip.json: replace api-client.ts ignore with domains/logs.ts

Closes #6693
