+++ turbo/knip.json
@@ -103,7 +103,7 @@
     "apps/platform/src/test/mocks/clerk-react-experimental.ts",
     "apps/web/src/lib/schedule/schedule-service.ts",
     "apps/web/src/lib/model-provider/model-provider-service.ts",
-    "apps/cli/src/lib/api/api-client.ts",
+    "apps/cli/src/lib/api/domains/logs.ts",
     "apps/cli/src/lib/utils/paginate.ts",
     "apps/cli/src/commands/cook/cook.ts",
     "apps/web/src/lib/auth/sandbox-token.ts",
